source/backend/requirements/base.txt
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
attrs==22.1.0
black==22.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.4
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
swagger-ui==0.1.2
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.3
whitenoise==6.2.0
